+++ EarlyBird.Packages.DAL/Models/Package.cs
@@ -2,7 +2,7 @@
 {
     public class Package
     {
-        public int Kolliid { get; set; }
+        public long Kolliid { get; set; }
         public double Weight { get; set; }
         public double Height { get; set; }
         public double Length { get; set; }
